another explicit fallthrough

+++ src/audio/load.c
@@ -64,7 +64,7 @@ Acmd* alAdpcmPull(void* filter, s16* outp, s32 outCount, s32 sampleOffset, Acmd*
     inp = AL_DECODER_IN;
     aLoadADPCM(ptr++, f->bookSize, K0_TO_PHYS(f->table->waveInfo.adpcmWave.book->book));
 
-    looped = (outCount + f->sample > (int)(f->loop.end)) && (f->loop.count != 0);
+    looped = ((int)(outCount + f->sample) > (int)(f->loop.end)) && (f->loop.count != 0);
     if (looped)
         nSam = f->loop.end - f->sample;
     else
